splitting out aws from main.go
